backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
itsdangerous==2.1.2
google-auth==2.26.1
requests==2.31.0
